refactor: rename api module to api-client
